Refactor: Restructure package layout
레이어 단위 패키지 구조 완성
