+++ yagu-route/src/main/java/org/teamtuna/yaguroute/aggregate/Team.java
@@ -1,0 +1,4 @@
+package org.teamtuna.yaguroute.aggregate;
+
+public class Team {
+}
