+++ yagu-route/src/main/java/org/teamtuna/yaguroute/controller/BookingController.java
@@ -1,4 +1,4 @@
-package org.teamtuna.yaguroute.controller.booking;
+package org.teamtuna.yaguroute.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.core.KafkaTemplate;
