+++ yagu-route/src/main/java/org/teamtuna/yaguroute/dto/TeamDTO.java
@@ -1,0 +1,4 @@
+package org.teamtuna.yaguroute.dto;
+
+public class TeamDTO {
+}
