+++ yagu-route/src/main/java/org/teamtuna/yaguroute/repository/TeamRepository.java
@@ -1,0 +1,4 @@
+package org.teamtuna.yaguroute.repository;
+
+public interface TeamRepository {
+}
